school/core/views.py
from django.shortcuts import render
from rest_framework import generics
from .models import Student,Course,Enrollment
from .serializers import StudentSerializer,CourseSerializer,StudentDetailSerializer,CourseDetailSerializer,EnrollmentSerializer

# Create your views here.
class StudentCreateView(generics.CreateAPIView):
    queryset = Student.objects.all()
    serializer_class = StudentSerializer


class CourseCreateView(generics.CreateAPIView):
    queryset = Course.objects.all()
    serializer_class = CourseSerializer


class EnrollmentStudentView(generics.CreateAPIView):
    queryset = Enrollment.objects.all()
    serializer_class = EnrollmentSerializer


class StudentDetailView(generics.RetrieveAPIView):
    queryset = Student.objects.all()
    serializer_class = StudentDetailSerializer


class CourseDetailView(generics.RetrieveAPIView):
    queryset = Course.objects.all()
    serializer_class = CourseDetailSerializer


